+++ packages/orq-rc/src/orq_ai_sdk/models/createcontactop.py
@@ -132,7 +132,7 @@ class CreateContactResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2026-03-06T09:14:08.503Z")
+    updated: Optional[datetime] = parse_datetime("2026-03-06T12:45:23.925Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")