+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -150,7 +150,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KK16TFASN6YHRVZK4YBPT69M"
+        "01KK1JX9KP10ES9R20YZJDYJH2"
     )
     r"""The unique identifier of the data source"""