+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KK16TFB0Q205RVC8RG32X7E8"
+        "01KK1JX9MDZW1VFP94TSX3091X"
     )
     r"""The unique identifier of the data source"""