+++ packages/orq-rc/src/orq_ai_sdk/models/updatetoolop.py
@@ -286,7 +286,7 @@ class RequestBodyTools(BaseModel):
 
     schema_: Annotated[UpdateToolRequestBodyToolsSchema, pydantic.Field(alias="schema")]
 
-    id: Optional[str] = "01KK16TF72WKBM6SR7EQC1P9BP"
+    id: Optional[str] = "01KK1JX9AA8BXZSBHJFKM5FF37"
 
     description: Optional[str] = None
 
@@ -1349,7 +1349,7 @@ class UpdateToolResponseBodyCodeExecutionTool(BaseModel):
     code_tool: UpdateToolResponseBodyCodeTool
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "tool_01KK16TF6XCH6HZXQKPWH2Q14J"
+        "tool_01KK1JX9A4ACJ6XN3AEXGRM47B"
     )
 
     display_name: Optional[str] = None
@@ -1480,7 +1480,7 @@ class UpdateToolResponseBodyTools(BaseModel):
         UpdateToolResponseBodyToolsSchema, pydantic.Field(alias="schema")
     ]
 
-    id: Optional[str] = "01KK16TF6WHGQRZT30A0Y3QGW6"
+    id: Optional[str] = "01KK1JX9A3KGHC1HKE3KQ4QV29"
 
     description: Optional[str] = None
 
@@ -1611,7 +1611,7 @@ class UpdateToolResponseBodyMCPTool(BaseModel):
     mcp: UpdateToolResponseBodyMcp
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "tool_01KK16TF6VA0G9KY7FWG31PE86"
+        "tool_01KK1JX9A2RBYKPMY6478VWFFR"
     )
 
     display_name: Optional[str] = None
@@ -1912,7 +1912,7 @@ class UpdateToolResponseBodyHTTPTool(BaseModel):
     http: UpdateToolResponseBodyHTTP
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "tool_01KK16TF6TR5DP98M2G6AF1K56"
+        "tool_01KK1JX9A0ED2T7T5F3Y7GM6JE"
     )
 
     display_name: Optional[str] = None
@@ -2107,7 +2107,7 @@ class UpdateToolResponseBodyJSONSchemaTool(BaseModel):
     json_schema: UpdateToolResponseBodyJSONSchema
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "tool_01KK16TF6S146RGFCDX45JYTBC"
+        "tool_01KK1JX99ZH0W8J3BVWNZM59DW"
     )
 
     display_name: Optional[str] = None
@@ -2306,7 +2306,7 @@ class UpdateToolResponseBodyFunctionTool(BaseModel):
     function: UpdateToolResponseBodyFunction
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "tool_01KK16TF6R8946S7CQ4TKR79T0"
+        "tool_01KK1JX99Y086JDQ9P5BRBE9T7"
     )
 
     display_name: Optional[str] = None