+++ packages/orq-rc/RELEASES.md
@@ -6339,3 +6339,13 @@ Based on:
 - [python v4.5.0-rc.36] packages/orq-rc
 ### Releases
 - [PyPI v4.5.0-rc.36] https://pypi.org/project/orq-ai-sdk/4.5.0-rc.36 - packages/orq-rc
+
+## 2026-03-06 12:48:14
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.745.0 (2.850.7) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.5.0-rc.37] packages/orq-rc
+### Releases
+- [PyPI v4.5.0-rc.37] https://pypi.org/project/orq-ai-sdk/4.5.0-rc.37 - packages/orq-rc